+++ include/BuyTowerScreen.hpp
@@ -3,8 +3,13 @@
 #include "Screen.hpp"
 
 class BuyTowerScreen : public Screen {
+private:
+  //event manager (used to register, deregister from events, and create them)
+  shared_ptr<EventManager> eventManager;
+  //Store the textLoader to make requests for strings and constants
+  shared_ptr<TextLoader> textLoader;
 public:
-  BuyTowerScreen(int windowX, int windowY);
+  BuyTowerScreen(shared_ptr<EventManager> eventManager,shared_ptr<TextLoader> textLoader,int windowX, int windowY);
 
   void draw(sf::RenderWindow& window);
 };