+++ include/MainMenuScreen.hpp
@@ -12,18 +12,22 @@ class MainMenuScreen : public Screen{
 private:
   int selectedItem;
   int numItems;
-  vector<sf::Text> screen;
-  sf::Font font;
+  vector<string> textStrings;
+//  sf::Text text;
+//  sf::Font font;
   //Store the textLoader to make requests for strings and constants
   shared_ptr<TextLoader> textLoader;
-
-  sf::Texture texture;
-  sf::Sprite sprite;
-
+  //event manager (used to register, deregister from events, and create them)
+  shared_ptr<EventManager> eventManager;
+  int check;
 public:
-  MainMenuScreen(int windowX, int windowY, int numItems,shared_ptr<TextLoader> textLoader);
+  MainMenuScreen(shared_ptr<EventManager> eventManager,int windowX, int windowY, int numItems,shared_ptr<TextLoader> textLoader);
   void initText();
   void draw(sf::RenderWindow& window);
   void moveUp();
   void moveDown();
+  void registerDelegates();
+  void handleKeyPress(const EventInterface& event);
+  void anyFunction();
+
 };