+++ include/Screen.hpp
@@ -6,11 +6,14 @@
 #include <memory>
 #include <functional>
 #include "TextLoader.hpp"
+#include "EventManager.hpp"
+#include "../include/Events/KeyPressEvent.hpp"
+#include "../include/Events/MousePressEvent.hpp"
 
 using std::cout;
 using std::string;
 using std::shared_ptr;
-
+using namespace std::placeholders;
 class Screen{
 protected:
   int windowX;