+++ src/BuyTowerScreen.cpp
@@ -1,10 +1,12 @@
 #include "BuyTowerScreen.hpp"
 
-BuyTowerScreen::BuyTowerScreen(int windowX, int windowY){
+BuyTowerScreen::BuyTowerScreen(shared_ptr<EventManager> eventManager,shared_ptr<TextLoader> textLoader,int windowX, int windowY){
   this -> windowX = windowX;
   this -> windowY = windowY;
+  this -> eventManager = eventManager;
+  this -> textLoader = textLoader;
 }
 
 void BuyTowerScreen::draw(sf::RenderWindow &window){
-  
+
 }