+++ src/EventManager.cpp
@@ -25,6 +25,8 @@ EventManager::~EventManager(){
 
 //add an event to the queue of registered events
 void EventManager::queueEvent(shared_ptr<EventInterface> event){
+//  cout << "queue event  " << endl;
+//  cout << register_queue -> size() << endl;
   //in case we do not find the event type we add it
   //to the map
   event_map::const_iterator it = EventDelegateMap.find(event -> getEventType());
@@ -38,10 +40,14 @@ void EventManager::queueEvent(shared_ptr<EventInterface> event){
 
 //add an event to the map of events and their delegates
 void EventManager::registerEvent(const EventType &type){
+
+//  cout << register_queue -> size() << endl;
   event_delegates newdelegateList;
   event_map::const_iterator it = EventDelegateMap.find(type);
 
   if(it == EventDelegateMap.end()){
+    cout << "register event " << endl;
+    cout << type.eventWithin << endl;
     EventDelegateMap.insert({type, newdelegateList});
   }
 
@@ -50,6 +56,9 @@ void EventManager::registerEvent(const EventType &type){
 //register a delegate to an event type. Think of this like a person (class) is subscribing
 // to a particular magazine (EvenetType) and is sending their address (EventDelegate)
 void EventManager::registerDelegate(const EventDelegate& d, const string &eventDelegateIdentifier,const EventType &type){
+  cout << "register delegate " << endl;
+  cout << eventDelegateIdentifier << endl;
+  //cout << register_queue -> size() << endl;
   //in case we do not find the event type we add it
   //to the map
   event_map::const_iterator it = EventDelegateMap.find(type);
@@ -62,6 +71,8 @@ void EventManager::registerDelegate(const EventDelegate& d, const string &eventD
   if(eventDelegatesEventMap.find(eventDelegateIdentifier) == eventDelegatesEventMap.end()){
     shared_ptr<EventDelegate> ed = make_shared<EventDelegate>(d);
     EventDelegateMap.at(type).insert({eventDelegateIdentifier,ed});
+    cout << EventDelegateMap.at(type).size() << endl;
+    cout << "delegate registered! " << endl;
   }
 
 }
@@ -78,6 +89,8 @@ void EventManager::deregisterDelegate(const EventDelegate& d, const string &even
 
 void EventManager::triggerEvent(const EventInterface& event)
 {
+//  cout << "trigger event " << endl;
+//  cout << register_queue -> size() << endl;
   //search to see if this event is registered
   auto delegateList = EventDelegateMap.find(event.getEventType());
   if(delegateList == EventDelegateMap.end()){
@@ -97,8 +110,10 @@ void EventManager::triggerEvent(const EventInterface& event)
 
 //iterate through all registered events and trigger them
 void EventManager::processEvent(){
-  std::swap(process_queue, register_queue);
-  clear(*register_queue);
+//  cout << "process  event " << endl;
+//  cout << register_queue -> size() << endl;
+  std::swap(*process_queue, *register_queue);
+  clear(register_queue);
   while(!process_queue->empty()){
     shared_ptr<EventInterface> event = process_queue->front();
     process_queue->pop();
@@ -107,29 +122,18 @@ void EventManager::processEvent(){
 
 }
 
-/*
- * search the passed list of event delegates for one and if it is not found
- * then return 0 (false), else return the index it is at (true)
- */
- /*
-int EventManager::find(event_delegates &eventDelegateList,  const EventDelegate& toFind){
-
-  int index = 0;
-  for(auto delegate: eventDelegateList){
-    if(&(*delegate) == &(toFind)){
-      return index;
-    }
-    index++;
-  }
-  return 0;
-}
-*/
 //used to clear the queue efficiently
-void EventManager::clear( event_queue &q )
+void EventManager::clear( shared_ptr<event_queue> q )
 {
-   if(q.size() != 0){
-     event_queue empty;
-     std::swap( q, empty );
+   if(q -> size() != 0){
+     q -> pop();
+     //event_queue empty;
+     //std::swap( *q, empty );
 
    }
+}
+
+
+void EventManager::dumpErrors(){
+  //cout << s.str();
 }