+++ src/PlayingScreen.cpp
@@ -1,9 +1,10 @@
 #include "PlayingScreen.hpp"
 
-PlayingScreen::PlayingScreen(int windowX, int windowY){
+PlayingScreen::PlayingScreen(shared_ptr<EventManager> eventManager,shared_ptr<TextLoader> textLoader,int windowX, int windowY){
   this -> windowX = windowX;
   this -> windowY = windowY;
-
+  this -> eventManager = eventManager;
+  this -> textLoader = textLoader;
 }
 
 void PlayingScreen::draw(sf::RenderWindow &window){